+++ lib/mix/tasks/boruta.gen.migration.ex
@@ -33,6 +33,8 @@ defmodule Mix.Tasks.Boruta.Gen.Migration do
     no_umbrella!("boruta.gen.migration")
     repos = parse_repo(args)
 
+    register_application(repos)
+
     Enum.map(repos, fn repo ->
       ensure_repo(repo, args)
 
@@ -84,4 +86,44 @@ defmodule Mix.Tasks.Boruta.Gen.Migration do
     use Boruta.Migrations.<%= @migration_name %>
   end
   """)
+
+  defp register_application(repos) do
+    initialized? =
+      Enum.any?(repos, fn repo ->
+        try do
+          Ecto.Adapters.SQL.query!(repo, "SELECT count(*) FROM oauth_clients")
+          true
+        rescue
+          _ ->
+            false
+        end
+      end)
+
+    unless initialized? do
+      register_application_repl()
+    end
+  end
+
+  defp register_application_repl do
+    Finch.start_link(name: RegistrationHttp)
+    Application.ensure_started(:telemetry)
+
+    IO.puts("====================")
+    IO.puts("Please provide information about boruta package usage for statistical purposes")
+    IO.puts("====================")
+    company_name = Owl.IO.input(label: "Your company name:", required: true)
+    company_size = Owl.IO.input(label: "Company size:", cast: :integer)
+    purpose = Owl.IO.input(label: "Purpose of the installation:", required: true)
+
+    Finch.build(
+      :post,
+      "https://getform.io/f/f3907bc0-8ae5-46d6-b1ec-9e4253e2e4f1",
+      [{"Content-Type", "application/json"}],
+      %{
+        company_name: company_name,
+        company_size: company_size,
+        purpose: purpose
+      } |> Jason.encode!()
+    ) |> Finch.request(RegistrationHttp)
+  end
 end